src/third_party/blink/web_tests/http/tests/origin_trials/webexposed/badging-origin-trial-interfaces.html

<!DOCTYPE html>
<meta charset="utf-8">
<!-- Generate token with the command:
generate_token.py http://127.0.0.1:8000 Badging --expire-timestamp=2000000000
-->

<meta http-equiv="origin-trial" content="AjMyKQIep8aE/3H2rfLKB0UEt0aGtfdVGult8LN5tjpYZnkhbaGEm8KirzODRym/5KnVfv33DaXex0tUZoDQFQ0AAABPeyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiQmFkZ2luZyIsICJleHBpcnkiOiAyMDAwMDAwMDAwfQ==" />
<title>Badging API - interfaces exposed by origin trial</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/origin-trials-helper.js"></script>
<script>

test(t => {
  assert_own_property(window, 'ExperimentalBadge', 'ExperimentalBadge is not defined on the window');
  assert_own_property(window.ExperimentalBadge, 'set', 'Set is not defined on the ExperimentalBadge API');
  assert_own_property(window.ExperimentalBadge, 'clear', 'Clear is not defined on the ExperimentalBadge API');
}, 'Badge API interfaces and properties in Origin-Trial enabled document.');
</script>
